+++ src/graphics/mod.rs
@@ -4,7 +4,6 @@ mod oam;
 pub mod palette;
 pub mod palette_memory;
 pub mod pattern_table;
-mod pixel_producer;
 pub mod ppu;
 mod ppu_registers;
 mod render_address;